refactor(waveform-seek): H3.5 — extract useWaveformHeights hook + hoist constants
